Fix PR CI test boundary failures
Move harness helper tests out of the stage unit suite and remove a duplicate prompt fixture key so full CI lint and Windows tests pass.
