+++ tests/test_skills_default_prompt_contract.py
@@ -164,7 +164,6 @@ async def test_default_prompt_only_injects_retained_bundled_skills(
                 "ffmpeg": True,
                 "ffprobe": True,
                 "xelatex": True,
-                "bibtex": True,
                 "nano-pdf": True,
                 "python": True,
                 "python3": True,
